First version of ww for Linux
This port for Linux support most of the functionalities of ww, including
workspace creation and activation. WW_DEFAULT_VOLUMES var was replaced by
the WW_DEFAULT_PATH in the linux version, since there is no "volume" concept
on Linux file systems.

Worth of noting that on Linux, workspace activation must be sourced (a
check was added to make sure the user "source" the activation).

EDEN-2221